switch to using floats for text positioning
git-svn-id: http://jmonkeyengine.googlecode.com/svn/trunk@3938 75d07b2b-3a1a-0410-a2c5-0572b91ccdca

+++ src/com/jme/renderer/jogl/JOGLFont.java
@@ -135,7 +135,7 @@ public class JOGLFont {
      * @param set
      *            the mode of font: NORMAL or ITALICS.
      */
-    public void print(JOGLRenderer r, int x, int y, Vector3f scale, StringBuffer text, int set) {
+    public void print(JOGLRenderer r, float x, float y, Vector3f scale, StringBuffer text, int set) {
         final GL gl = GLU.getCurrentGL();
 
         RendererRecord matRecord = (RendererRecord) DisplaySystem.getDisplaySystem().getCurrentContext().getRendererRecord();
